Move sha256.c to lib/
Precompiled headers don't work if we're not a pure C++ project.
