+++ lib/sha256/cmake_config.h.in
@@ -1,0 +1,5 @@
+// Filled in by the build system
+
+#pragma once
+
+#cmakedefine HAVE_ENDIAN_H
